Software Construction - OODesign/robust-classes-practice-problems/FoodTruck/src/model/Owner.java
package model;

public class Owner {

    private String name;
    private Kitchen kitchen;

    public Owner(String name, Kitchen kit) {
        this.name = name;
        kitchen = kit;
    }

    //getters
    public String getName() { return name; }
    public Kitchen getKitchen() { return kitchen; }

    public boolean orderMoreTacos(int amount) {
        try {
            kitchen.makeTaco(amount);
            return true;
        } catch (Exception e) {
            System.out.println(e.getMessage());
            return false;
        }
    }

    public boolean askForMoreIngredients(int amount) {
        try {
            kitchen.buyIngredients(amount);
            return true;
        } catch (Exception e) {
            System.out.println(e.getMessage());
            return false;
        }
    }
}
